+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# v0.11.17
+- Added user id parameter in get entities hashes endpoint.
+- Added user id parameter in validate entities data endpoint.
+
 # v0.11.16
 - Fixed validate entity hashes with entity granted.